move utils class and method out.
